Move convert functions to new convert module
