+++ iodata/formats/cp2klog.py
@@ -23,7 +23,8 @@ from typing import Union
 import numpy as np
 from numpy.typing import NDArray
 
-from ..basis import HORTON2_CONVENTIONS, MolecularBasis, Shell, angmom_sti
+from ..basis import MolecularBasis, Shell, angmom_sti
+from ..convert import HORTON2_CONVENTIONS
 from ..docstrings import document_load_one
 from ..orbitals import MolecularOrbitals
 from ..overlap import factorial2
