+++ iodata/formats/molden.py
@@ -33,14 +33,8 @@ import attrs
 import numpy as np
 from numpy.typing import NDArray
 
-from ..basis import (
-    HORTON2_CONVENTIONS,
-    MolecularBasis,
-    Shell,
-    angmom_its,
-    angmom_sti,
-    convert_conventions,
-)
+from ..basis import MolecularBasis, Shell, angmom_its, angmom_sti
+from ..convert import HORTON2_CONVENTIONS, convert_conventions
 from ..docstrings import document_dump_one, document_load_one
 from ..iodata import IOData
 from ..orbitals import MolecularOrbitals
