+++ iodata/formats/molekel.py
@@ -29,7 +29,8 @@ from warnings import warn
 import numpy as np
 from numpy.typing import NDArray
 
-from ..basis import MolecularBasis, Shell, angmom_its, angmom_sti, convert_conventions
+from ..basis import MolecularBasis, Shell, angmom_its, angmom_sti
+from ..convert import convert_conventions
 from ..docstrings import document_dump_one, document_load_one
 from ..iodata import IOData
 from ..orbitals import MolecularOrbitals
