+++ iodata/formats/wfn.py
@@ -32,7 +32,8 @@ from typing import TextIO
 import numpy as np
 from numpy.typing import NDArray
 
-from ..basis import MolecularBasis, Shell, convert_conventions
+from ..basis import MolecularBasis, Shell
+from ..convert import convert_conventions
 from ..docstrings import document_dump_one, document_load_one
 from ..iodata import IOData
 from ..orbitals import MolecularOrbitals
